requirements.txt
yfinance
transformers
torch
beautifulsoup4
python-telegram-bot==13.15